windows/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.5.3
beautifulsoup4==4.12.2
lxml==4.9.3
